+++ test/typescript_test.ts
@@ -441,7 +441,7 @@ describe('TypeScript support', () => {
     }
 
     interface BookDb {
-      books: Automerge.Table<Book, ['authors', 'title', 'isbn']>
+      books: Automerge.Table<Book>
     }
 
     // Example data
@@ -457,56 +457,47 @@ describe('TypeScript support', () => {
     }
 
     let s1: Doc<BookDb>
-    let id: string
+    let id: Automerge.UUID
+    let ddiaWithId: Book & Automerge.TableRow
 
     beforeEach(() => {
       s1 = Automerge.change(Automerge.init<BookDb>(), doc => {
-        doc.books = new Automerge.Table(['authors', 'title', 'isbn'])
+        doc.books = new Automerge.Table()
         id = doc.books.add(DDIA)
       })
+      ddiaWithId = Object.assign({id}, DDIA)
     })
 
-    it('supports `byId`', () => assert.deepStrictEqual(s1.books.byId(id), DDIA))
+    it('supports `byId`', () => assert.deepStrictEqual(s1.books.byId(id), ddiaWithId))
     it('supports `count`', () => assert.strictEqual(s1.books.count, 1))
     it('supports `ids`', () => assert.deepStrictEqual(s1.books.ids, [id]))
-    it('supports iteration', () => assert.deepStrictEqual([...s1.books], [DDIA]))
-    it('supports `columns`', () => assert.deepStrictEqual(s1.books.columns, ['authors', 'title', 'isbn']))
+    it('supports iteration', () => assert.deepStrictEqual([...s1.books], [ddiaWithId]))
 
-    it('allows modifying the columns array', () => {
-      const s2 = Automerge.change(s1, doc => doc.books.columns.push('publisher'))
-      assert.deepStrictEqual(s2.books.columns, ['authors', 'title', 'isbn', 'publisher'])
-      assert.deepStrictEqual(s2.books.byId(id), DDIA)
-
+    it('allows adding row properties', () => {
       // Note that if we add columns and want to actually use them, we need to recast the table to a
       // new type e.g. without the `ts-ignore` flag, this would throw a type error:
-      // @ts-ignore
-      const p2 = s2.books.byId(id).publisher // Property 'publisher' does not exist on type book
 
-      // So we need to create new types:
+      // @ts-ignore - Property 'publisher' does not exist on type book
+      const p2 = s1.books.byId(id).publisher 
+
+      // So we need to create new types
       interface BookDeluxe extends Book {
+        // ... existing properties, plus:
         publisher?: string
       }
       interface BookDeluxeDb {
-        books: Automerge.Table<BookDeluxe, ['authors', 'title', 'isbn', 'publisher']>
+        books: Automerge.Table<BookDeluxe>
       }
 
+      const s2 = s1 as Doc<BookDeluxeDb> // Cast existing table to new type
       const s3 = Automerge.change(
-        s2 as BookDeluxeDb, // Cast existing table to new type
+        s2,
         doc => (doc.books.byId(id).publisher = "O'Reilly")
       )
+
       // Now we're off to the races
       const p3 = s3.books.byId(id).publisher
-      assert.deepStrictEqual(p3, "O'Reilly")
-
-      // and we can even do this:
-      Automerge.change(s3, doc => {
-        doc.books.add([
-          ['Cachin, Christian', 'Guerraoui, Rachid', 'Rodrigues, Luís'],
-          'Introduction to Reliable and Secure Distributed Programming',
-          '3-642-15259-7',
-          'Springer',
-        ])
-      })
+      assert.strictEqual(p3, "O'Reilly")
     })
 
     it('supports `remove`', () => {
@@ -515,30 +506,19 @@ describe('TypeScript support', () => {
     })
 
     describe('supports `add`', () => {
-      it('accepts value passed as correctly-ordered array', () => {
-        let bookId: string
-        const s2 = Automerge.change(s1, doc => {
-          bookId = doc.books.add([
-            ['Cachin, Christian', 'Guerraoui, Rachid', 'Rodrigues, Luís'],
-            'Introduction to Reliable and Secure Distributed Programming',
-            '3-642-15259-7',
-          ])
-        })
-        assert.deepStrictEqual(s2.books.byId(bookId), RSDP)
-      })
-
       it('accepts value passed as object', () => {
         let bookId: string
         const s2 = Automerge.change(s1, doc => (bookId = doc.books.add(RSDP)))
-        assert.deepStrictEqual(s2.books.byId(bookId), RSDP)
+        assert.deepStrictEqual(s2.books.byId(bookId), Object.assign({id: bookId}, RSDP))
+        assert.strictEqual(s2.books.byId(bookId).id, bookId)
       })
     })
 
     describe('standard array operations on rows', () => {
       it('supports `filter`', () =>
-        assert.deepStrictEqual(s1.books.filter(book => book.authors.length === 1), [DDIA]))
+        assert.deepStrictEqual(s1.books.filter(book => book.authors.length === 1), [ddiaWithId]))
       it('supports `find`', () =>
-        assert.deepStrictEqual(s1.books.find(book => book.isbn === '1449373321'), DDIA))
+        assert.deepStrictEqual(s1.books.find(book => book.isbn === '1449373321'), ddiaWithId))
       it('supports `map`', () =>
         assert.deepStrictEqual(s1.books.map<string>(book => book.title), [DDIA.title]))
     })