+++ CHANGELOG.md
@@ -20,6 +20,21 @@ is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.14.0] — 2020-03-25
+
+- **Removed** [#236]: Undocumented `Automerge.Table` API that allowed rows to be added by
+  providing an array of values. Now rows must be given as an object ([@HerbCaudill])
+- **Removed** [#241]: Constructor of `Automerge.Table` no longer takes an array of columns, and
+  the `columns` property of `Automerge.Table` is also removed ([@ept])
+- **Changed** [#242]: Rows of `Automerge.Table` now automatically get an `id` property containing
+  the primary key of that row ([@ept])
+- **Removed** [#243]: `Automerge.Table` objects no longer have a `set()` method. Use `add()` or
+  `remove()` instead ([@ept])
+- **Removed** support for Node 8, which is no longer being maintained
+- **Added** [#194], [#238]: `Automerge.Text` objects may now contain objects as well as strings;
+  new method `Text.toSpans()` that concatenates characters while leaving objects unchanged
+  ([@pvh], [@ept], [@nornagon])
+
 ## [0.13.0] — 2020-02-24
 
 - **Added** [#232]: New API `Automerge.getAllChanges()` returns all changes ([@ept])
@@ -230,7 +245,8 @@ is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - First public release.
 
 
-[Unreleased]: https://github.com/automerge/automerge/compare/v0.13.0...HEAD
+[Unreleased]: https://github.com/automerge/automerge/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/automerge/automerge/compare/v0.13.1...v0.14.0
 [0.13.0]: https://github.com/automerge/automerge/compare/v0.12.1...v0.13.0
 [0.12.1]: https://github.com/automerge/automerge/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/automerge/automerge/compare/v0.11.0...v0.12.0
@@ -259,12 +275,18 @@ is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 [0.4.0]: https://github.com/automerge/automerge/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/automerge/automerge/compare/v0.2.0...v0.3.0
 
+[#243]: https://github.com/automerge/automerge/pull/243
+[#242]: https://github.com/automerge/automerge/pull/242
+[#241]: https://github.com/automerge/automerge/pull/241
+[#238]: https://github.com/automerge/automerge/pull/238
+[#236]: https://github.com/automerge/automerge/pull/236
 [#232]: https://github.com/automerge/automerge/pull/232
 [#230]: https://github.com/automerge/automerge/issues/230
 [#219]: https://github.com/automerge/automerge/issues/219
 [#210]: https://github.com/automerge/automerge/pull/210
 [#202]: https://github.com/automerge/automerge/issues/202
 [#199]: https://github.com/automerge/automerge/pull/199
+[#194]: https://github.com/automerge/automerge/issues/194
 [#184]: https://github.com/automerge/automerge/pull/184
 [#183]: https://github.com/automerge/automerge/pull/183
 [#181]: https://github.com/automerge/automerge/pull/181
@@ -339,6 +361,7 @@ is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 [@minhhien1996]: https://github.com/minhhien1996
 [@mmcgrana]: https://github.com/mmcgrana
 [@mmmm1998]: https://github.com/mmmm1998
+[@nornagon]: https://github.com/nornagon
 [@pvh]: https://github.com/pvh
 [@salzhrani]: https://github.com/salzhrani
 [@saranrapjs]: https://github.com/saranrapjs